requirements.txt
flask==3.0.2
openai==1.3.0
python-dotenv==1.0.1
werkzeug==3.0.1
sentence-transformers==2.5.1
faiss-cpu==1.7.4
numpy==1.26.4
markdown==3.5.2
httpx==0.24.1
urllib3<2.0.0 